4.3.3
- Updated `SerializeOrderAttribute` to cast `position` to `Int32`.
- Added new `using` directives in `FormatterProvider.cs` for serialization namespaces.
- Enhanced `Get<T>()` method to skip null marker support for `SerializePackableAttribute`.
- Introduced debug logging in `UnmanagedFormatter<T>.Serialize` for better traceability.
- Added `using` directive for `Nalix.Shared.Messaging.Binary` in `FieldCache.Core.cs`.
- Implemented debug logging in `FieldCache<T>` static constructor for layout insights.
- Corrected field population in `DiscoverFields<TField>()` method.
- Modified property order retrieval to include non-public properties.
- Updated assertions in `ConnectionLimiterTests.cs` for accurate validation.
- Changed configuration in `RequestLimiterTests.cs` from `RateLimitOptions` to `TokenBucketOptions`.
- Adjusted `Binary256Tests.cs` to instantiate `Binary128` correctly.
- Added new tests in `LiteSerializer_Class_Tests.cs` for `Binary128` serialization and deserialization.

+++ src/Nalix.Common/Serialization/Attributes/SerializeOrderAttribute.cs
@@ -27,5 +27,5 @@ public class SerializeOrderAttribute : System.Attribute
     /// Initializes a new instance of the <see cref="SerializeOrderAttribute"/> class using an enum value for serialization order.
     /// </summary>
     /// <param name="position">The enum value that defines the order of serialization.</param>
-    public SerializeOrderAttribute(PacketHeaderOffset position) : this((System.Byte)position) { }
+    public SerializeOrderAttribute(PacketHeaderOffset position) : this((System.Int32)position) { }
 }
